finish remaining routes, need to clean up logging
