+++ lib/utilities.js
@@ -14,6 +14,7 @@ const DropCache = require('./caches/DropCache.js');
 const Cache = require('./caches/cache.js');
 
 const platforms = ['pc', 'ps4', 'xb1', 'swi'];
+const platformAliases = ['ns'];
 const worldStates = {};
 const parser = function parser(data) {
   return new Worldstate(data);
@@ -48,19 +49,20 @@ delete warframeData.weapons;
 delete warframeData.warframes;
 
 const wfKeys = Object.keys(warframeData).map(key => key.toLowerCase());
-wfKeys.push('drops', 'mods', 'pc', 'ps4', 'xb1', 'swi', 'heartbeat', 'warframe');
+wfKeys.push('drops', 'mods', 'pc', 'ps4', 'xb1', 'swi', 'ns', 'heartbeat', 'warframe');
 
 module.exports = {
   setHeadersAndJson: (res, json) => {
     res.setHeader('Content-Type', 'application/json');
     res.setHeader('Access-Control-Allow-Origin', '*');
-    res.setHeader('Access-Control-Allow-Methods', 'GET, POST, DELETE, OPTIONS');
+    res.setHeader('Access-Control-Allow-Methods', 'GET, OPTIONS');
     res.setHeader('Access-Control-Allow-Headers', 'DNT,X-CustomHeader,Keep-Alive,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Content-Range,Range');
     res.setHeader('Access-Control-Expose-Headers', 'DNT,X-CustomHeader,Keep-Alive,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Content-Range,Range');
     res.json(json);
   },
   logger,
-  platforms: ['pc', 'ps4', 'xb1', 'swi'],
+  platforms,
+  platformAliases,
   cache: apiCache.middleware,
   router: express.Router(),
   Items,
